requirements.txt
apache-airflow-providers-apache-spark
requests